[fix]Account bug in worker
